Added dual-boot module
Helps my machines go to and from Windows/Ubuntu

dual-boot:
./readme.first the readme file
notes on using, prerequisites and a basic summary (from memory)
on how to get a FAT32 /boot partition for grub.

./ubuntu/windows goes from ubuntu => windows
Unfortunately we can't deterministically figure out what position
the windows bootloader is in (at the very least I don't know how
Grub figures this position out with the hiding that's going on).

./windows/ubu.bat goes from windows => ubuntu
Takes grub's settings and puts default=0, which means
"Load Ubuntu", then restarts the system.

./windows/sed.exe
like sed in linux, but for Windows. Tried to figure out how to do
this automatically with just command line implements but couldn't
figure anything out in that direction in a reasonable timeframe